% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-rules.R
\name{ks_collapse_check_msgs}
\alias{ks_collapse_check_msgs}
\title{Collapse list-column of check messages to a character column}
\usage{
ks_collapse_check_msgs(data, sep = "; ", col)
}
\arguments{
\item{data}{A data frame containing a list-column of check messages.}

\item{sep}{Character scalar. Separator used when collapsing multiple
messages. Default \code{"; "}.}

\item{col}{\code{<tidy-select>} Bare name of the list-column to collapse.
Must be provided explicitly.}
}
\value{
\code{data} with \code{col} replaced by a character column.
}
\description{
Converts the list-column produced by \code{\link{ks_check_rules}} into a
plain character column by pasting multiple messages together with a
separator.
}
\examples{
\dontrun{
df_checked <- ks_check_rules(data, rules, mode = "all")
ks_collapse_check_msgs(df_checked, col = check_msgs)
}

}
